refactor: 당첨 등수 계산 책임 이동(WinningLotto -> Lottos)
